+++ parrot/src/buffers/index.rs
@@ -1,4 +1,3 @@
-use super::Buffer;
 /// Represents the index buffer
 #[derive(Debug)]
 pub struct IndexBuffer {
@@ -12,11 +11,5 @@ impl IndexBuffer {
     pub fn slice(&self) -> wgpu::BufferSlice {
         self.wgpu
             .slice(0..(self.elements as usize * std::mem::size_of::<u16>()) as u64)
-    }
-}
-
-impl Buffer for IndexBuffer {
-    fn wgpu_buffer(&self) -> &wgpu::Buffer {
-        &self.wgpu
     }
 }