+++ pigeon/src/pipeline/mod.rs
@@ -1,5 +1,7 @@
-pub mod triangle;
 use pigeon_parrot as parrot;
 use parrot::{
     pipeline::PipelineLayout,
 };
+
+pub mod triangle;
+pub mod egui;